+++ docs/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [??] - ??
 
+### Added
+
+- Added support for the `--draft` option when deploying content,
+  this allows to deploy a new bundle for the content without exposing
+  it as a the activated one.
+
 ## [1.26.0] - 2025-05-28
 
 ### Added